+++ src/config/parameter.cpp
@@ -53,7 +53,7 @@ void parameter::initialize(const po::option_description& option,
 int parameter::position(const po::option_description& option,
     const argument_list& arguments) const
 {
-    return find_pair_position(arguments, option.long_name());
+    return static_cast<int>(find_pair_position(arguments, option.long_name()));
 }
 
 // 100% unit coverage, all three scenarios (long, short, both)
